Admin Page: display update modal correctly
Use a new state index, 'display_update_modal' to indicate whether the update modal
should display. This state index will hold a boolean, and the value will be true
when the update modal should be displayed.

In Jetpack_React_Page::get_initial_state(), use state('display_update_modal') to
determine whether to collect the release post content and send it to the admin
page.

Move the update modal methods from the Jetpack class to the Jetpack_React_Page class.
They're only used in the Jetpack_React_Page class, so they should be located there.

These methods must be moved and should no longer be static:
 - Jetpack::send_update_modal_data() -> change to Jetpack_React_Page::get_update_modal_data()
 - Jetpack::allow_post_embed_iframe() (same method name)
 - Jetpack::get_release_post_data() (same method name)

To prevent the modal from being displayed twice if the Jetpack dashboard is visited immediately
after updating Jetpack, only set the jetpackState['display_update_modal'] value in the cookie
if the current screen is not the Jetpack dashboard.